requirements.txt
# requirements.txt
Flask
Flask-Cors
requests
python-whois
dnspython
webtech
gunicorn